Implemented Save/Load functions, Created header file
The save and load functions are now created and work. They will save
and load to a file called “inventory.txt” either a Card entry per line,
and read in order. Will soon convert this to a binary R/W file. Created
a makefile to make things easy. Also created a header file for the
PFModus.