User input loop finished
Starting at the sylladexFramework file, it takes user input to enter a
modus file which begins a user input loop to obtain feedback on which
functions to perform.
Current bug: strings obtained from fgets causes the output to create a
newline at the end of strings placed with %s in fprint.