Updated Modus, moved Metadata to model package, updated ModusManager
ModusManager can now add all modi in the modus package to it’s
modusList in a modular fashion (i.e. modus classes don’t need to be
hardcoded into the ModusManager to be added to the view).  Modus got
some javaDoc update and also added a function to make a Metadata field
necessary for all subclasses. Sylladex had more progress made to it.