Overhauled architecture
created ModusBuffer to act as a buffer that contains all the required data a modus would need to access and support modus hot swaps using the inverted dependency technique. sylladex was refactored to not share static objects and instead pass wrapped objects to the modus buffer on modus manager construction. sylladex is pruned of unnecessary dependencies. modus manager now handles modus input and acts as an API for the modus to the sylladex using the facade pattern technique. Modus now relies on the modus buffer for data from the sylladex, using the modus buffer's public getters.